refactor: setup service project
Prepare to transfer responsibilities from the Kotlin project to a .NET
project. The current thinking is that gRPC will be the method of
communication between the bot and the service.

Upgrade the existing bot to the lastest version of Kotlin & Kord.
