+++ bot/Dockerfile
