+++ bot/gradle.properties
