+++ bot/settings.gradle.kts
