+++ bot/src/main/kotlin/com/guigou/botisgud/extensions/java/util/Optional.kt
