+++ bot/src/main/kotlin/com/guigou/botisgud/extensions/kord/ReactionAddEvent.kt
@@ -1,0 +1,7 @@
+package com.guigou.botisgud.extensions.kord
+
+import dev.kord.core.event.message.ReactionAddEvent
+import io.ktor.http.Url
+
+val ReactionAddEvent.link: Url
+    get() = Url("https://discordapp.com/channels/${guildId?.value ?: "@me"}/${channelId.value}/${messageId.value}")
