+++ bot/src/main/kotlin/com/guigou/botisgud/models/Reminder.kt
