+++ bot/src/main/kotlin/com/guigou/botisgud/models/ReminderTrigger.kt
