+++ bot/src/main/kotlin/com/guigou/botisgud/models/User.kt
