+++ bot/src/main/kotlin/com/guigou/botisgud/services/nickname/NicknameEntity.kt
