+++ bot/src/main/kotlin/com/guigou/botisgud/services/nickname/NicknameService.kt
@@ -1,0 +1,7 @@
+package com.guigou.botisgud.services.nickname
+
+import dev.kord.common.entity.Snowflake
+
+interface NicknameService {
+    suspend fun getUsersByGuild(): Map<Snowflake, List<Snowflake>>
+}
