+++ bot/src/main/kotlin/com/guigou/botisgud/services/user/UserEntity.kt
