+++ bot/src/main/kotlin/com/guigou/botisgud/services/user/UserService.kt
@@ -1,0 +1,9 @@
+package com.guigou.botisgud.services.user
+
+import dev.kord.common.entity.Snowflake
+import com.guigou.botisgud.models.User
+
+interface UserService {
+    suspend fun add(user: User)
+    suspend fun get(userId: Snowflake): User?
+}
