+++ bot/src/main/resources/simplelogger.properties
