--- src/main/kotlin/com/guigou/botisgud/services/nickname/NicknameService.kt
@@ -1,7 +1,0 @@
-package com.guigou.botisgud.services.nickname
-
-import com.gitlab.kordlib.common.entity.Snowflake
-
-interface NicknameService {
-    suspend fun getUsersByGuild(): Map<Snowflake, List<Snowflake>>
-}
