--- src/main/kotlin/com/guigou/botisgud/services/user/UserService.kt
@@ -1,9 +1,0 @@
-package com.guigou.botisgud.services.user
-
-import com.gitlab.kordlib.common.entity.Snowflake
-import com.guigou.botisgud.models.User
-
-interface UserService {
-    suspend fun add(user: User)
-    suspend fun get(userId: Snowflake): User?
-}
